requirements.txt
fastapi[all]==0.98.0
transformers==4.31.0
datasets==2.14.3
Pillow==9.5.0
redis==4.6.0
gunicorn==20.1.0